+++ .github/workflows/android-config-prod/MainActivity.java
@@ -2,11 +2,13 @@ package com.optimize.elykia;
 
 import android.os.Bundle;
 import com.getcapacitor.BridgeActivity;
+import com.getcapacitor.community.database.sqlite.CapacitorSQLitePlugin;
 
 public class MainActivity extends BridgeActivity {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         registerPlugin(AppUpdatePlugin.class);
+        registerPlugin(CapacitorSQLitePlugin.class);
         super.onCreate(savedInstanceState);
     }
 }